applied openings api combined
