+++ app/controllers/app/interviewee.js
@@ -46,49 +46,7 @@ function getTotalExperience(req, res, next) {
   });
 }
 
-/**
- * Function to update Secondary Details
- * @param {Object} req Request Object
- * @param {Object} res Response Object
- * @param {Function} next Function to pass control to the next middleware
- */
-function updateSecondaryDetails(req, res, next) {
-  appIntervieweeLib.updateSecondaryOfInterviewee(
-    req.user._id,
-    req.body.secondary,
-    function(errInUpdation, updatedInstance) {
-      if (errInUpdation) {
-        res.status(500).json(errInUpdation);
-      } else {
-        res.status(200).json(updatedInstance);
-      }
-    }
-  );
-}
-
-/**
- * Function to update Degree Details
- * @param {Object} req Request Object
- * @param {Object} res Response Object
- * @param {Function} next Function to pass control to the next middleware
- */
-function updateDegreeDetails(req, res, next) {
-  appIntervieweeLib.updateDegreeOfInterviewee(
-    req.user._id,
-    req.body.degree,
-    function(errInUpdation, updatedInstance) {
-      if (errInUpdation) {
-        res.status(500).json(errInUpdation);
-      } else {
-        res.status(200).json(updatedInstance);
-      }
-    }
-  );
-}
-
 router.post('/', registerNewInterviewee);
 router.get('/details', getTotalExperience);
-router.post('/updateSecondaryDetails', updateSecondaryDetails);
-router.post('/updateDegreeDetails', updateDegreeDetails);
 
 module.exports = router;
