+++ app/controllers/interviewee/index.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const openingRoutes = require('./opening');
+const profileRoutes = require('./profile');
 const CONSTANTS = require('../../../config/constants');
 
 const router = express.Router();
@@ -17,4 +18,5 @@ function authenticator(req, res, next) {
 
 router.use(authenticator);
 router.use('/opening', openingRoutes);
+router.use('/profile', profileRoutes);
 module.exports = router;
