+++ app/controllers/interviewee/opening.js
@@ -39,32 +39,13 @@ function eligibleOpenings(req, res, next) {
 }
 
 /**
- * Function to fetch all applied upcoming openings
- * @param  {Object}   req  Request Object
- * @param  {Object}   res  Response Object
- * @param  {Function} next Function to pass control to the next middleware
- */
-function upcomingAppliedOpenings(req, res, next) {
-  openingLib.upcomingAppliedOpenings(req.user._id, function(
-    err,
-    fetchedAppliedOpenings
-  ) {
-    if (err) {
-      res.status(500).json(err);
-      return;
-    }
-    res.status(200).json(fetchedAppliedOpenings);
-  });
-}
-
-/**
  * Function to fetch all applied openings which are to be given today
  * @param  {Object}   req  Request Object
  * @param  {Object}   res  Response Object
  * @param  {Function} next Function to pass control to the next middleware
  */
-function currentAppliedOpenings(req, res, next) {
-  openingLib.currentAppliedOpenings(req.user._id, function(
+function appliedOpenings(req, res, next) {
+  openingLib.appliedOpenings(req.user._id, function(
     err,
     fetchedAppliedOpenings
   ) {
@@ -78,6 +59,5 @@ function currentAppliedOpenings(req, res, next) {
 
 router.post('/apply', applyForOpening);
 router.get('/eligible', eligibleOpenings);
-router.get('/upcomingApplied', upcomingAppliedOpenings);
-router.get('/currentApplied', currentAppliedOpenings);
+router.get('/applied', appliedOpenings);
 module.exports = router;
