+++ lib/interviewee/opening.js
@@ -162,32 +162,6 @@ function extractUpcomingApplied(tracks, callback) {
 }
 
 /**
- * Function to first getting all interviewee tracks, further fetching applied openings
- * @param {ObjectId} userId
- * @param {Function} callback having error and data
- */
-function upcomingAppliedOpenings(userId, callback) {
-  async.waterfall(
-    [
-      function(waterfallCallback) {
-        getAllIntervieweeTracks(userId, function(err, tracks) {
-          waterfallCallback(err, tracks);
-        });
-      },
-
-      function(tracks, waterfallCallback) {
-        extractUpcomingApplied(tracks, function(err, Opening) {
-          waterfallCallback(err, Opening);
-        });
-      },
-    ],
-    function(err, data) {
-      callback(err, data);
-    }
-  );
-}
-
-/**
  * Function to fetch applied interviews which are to be taken on current date
  * @param {Object} tracks
  * @param {Function} callback having error and fetched data
@@ -220,11 +194,11 @@ function extractCurrentApplied(tracks, callback) {
 }
 
 /**
- * Function to first getting all interviewee tracks, further fetching current applied openings
+ * Function for getting upcoming applied opening and today's applied opening
  * @param {ObjectId} userId
  * @param {Function} callback having error and data
  */
-function currentAppliedOpenings(userId, callback) {
+function appliedOpenings(userId, callback) {
   async.waterfall(
     [
       function(waterfallCallback) {
@@ -234,9 +208,23 @@ function currentAppliedOpenings(userId, callback) {
       },
 
       function(tracks, waterfallCallback) {
-        extractCurrentApplied(tracks, function(err, Opening) {
-          waterfallCallback(err, Opening);
+        extractUpcomingApplied(tracks, function(err, upcomingOpening) {
+          waterfallCallback(err, tracks, upcomingOpening);
         });
+      },
+
+      function(tracks, upcomingOpening, waterfallCallback) {
+        extractCurrentApplied(tracks, function(err, currentOpening) {
+          waterfallCallback(err, upcomingOpening, currentOpening);
+        });
+      },
+
+      function(upcomingOpening, currentOpening, waterfallCallback) {
+        const applied = {
+          currentAppliedOpenings: currentOpening,
+          upcomingAppliedOpenings: upcomingOpening,
+        };
+        waterfallCallback(null, applied);
       },
     ],
     function(err, data) {
@@ -369,6 +357,5 @@ function applyForOpening(openingId, userId, callback) {
 module.exports = {
   applyForOpening: applyForOpening,
   eligibleOpenings: eligibleOpenings,
-  upcomingAppliedOpenings: upcomingAppliedOpenings,
-  currentAppliedOpenings: currentAppliedOpenings,
+  appliedOpenings: appliedOpenings,
 };
